+++ pkgs/dart_mcp/lib/src/client/client.dart
@@ -29,7 +29,7 @@ part 'sampling_support.dart';
 /// this will be invoked at the end of base class constructor.
 base class MCPClient {
   /// A description of the client sent to servers during initialization.
-  final ClientImplementation implementation;
+  final Implementation implementation;
 
   MCPClient(this.implementation) {
     initialize();
@@ -128,10 +128,10 @@ base class ServerConnection extends MCPBase {
   /// protocol version.
   late ProtocolVersion protocolVersion;
 
-  /// The [ServerImplementation] returned from the [initialize] request.
+  /// The [Implementation] returned from the [initialize] request.
   ///
   /// Only non-null after [initialize] has successfully completed.
-  ServerImplementation? serverInfo;
+  Implementation? serverInfo;
 
   /// The [ServerCapabilities] returned from the [initialize] request.
   ///