+++ pkgs/dart_mcp/test/test_utils.dart
@@ -78,16 +78,13 @@ class TestEnvironment<Client extends MCPClient, Server extends MCPServer> {
 
 base class TestMCPClient extends MCPClient {
   TestMCPClient()
-    : super(ClientImplementation(name: 'test client', version: '0.1.0'));
+    : super(Implementation(name: 'test client', version: '0.1.0'));
 }
 
 base class TestMCPServer extends MCPServer {
   TestMCPServer(super.channel, {super.protocolLogSink})
     : super.fromStreamChannel(
-        implementation: ServerImplementation(
-          name: 'test server',
-          version: '0.1.0',
-        ),
+        implementation: Implementation(name: 'test server', version: '0.1.0'),
         instructions: 'A test server',
       );
 }