+++ pkgs/dart_mcp_server/lib/src/server.dart
@@ -40,7 +40,7 @@ final class DartMCPServer extends MCPServer
     this.forceRootsFallback = false,
     super.protocolLogSink,
   }) : super.fromStreamChannel(
-         implementation: ServerImplementation(
+         implementation: Implementation(
            name: 'dart and flutter tooling',
            version: '0.1.0-wip',
          ),