+++ pkgs/dart_mcp_server/test/test_harness.dart
@@ -282,7 +282,7 @@ final class AppDebugSession {
 final class DartToolingMCPClient extends MCPClient with RootsSupport {
   DartToolingMCPClient()
     : super(
-        ClientImplementation(
+        Implementation(
           name: 'test client for the dart tooling mcp server',
           version: '0.1.0',
         ),